backend/requirements.txt
# IROAS BOSS V2 Backend Dependencies
# Phase 21 Authentication System Requirements

# FastAPI Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
pyotp==2.9.0
qrcode[pil]==7.4.2

# Data Validation
pydantic[email]==2.5.0

# HTTP Client
httpx==0.25.2

# Utilities
python-dotenv==1.0.0
python-dateutil==2.8.2
pytz==2023.3

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# MLM Business Requirements
pandas==2.1.4
openpyxl==3.1.2
chardet==5.2.0